feat(schematics): move command-cli to a separated package
As command-line nx isn't, stickly speaking, part of the schematics, it seems logical to move it to
a new, separated ``@nrwl/cli`` package.

This will permit, to begin with, two things :
1. use the nx cli in any project : in order to permit this, we'll need to add some more
flexibility, for example by permiting to configure the folders / projects architecture
2. make the nx cli compatible with angular-cli v6 before the global nrwl/nx refactoring, as it
could be done really faster

The next step could be to make a PR to angular/devkit or angular-cli in order to add some of the
features covered by the cli (mainly "affected") to it.

For now, the e2e tests are failing, as the addition of the @nrwl/cli dependency to the package.json
schematics will require a first release of this new package.

NB: as ``@nrwl/schematics`` requires ``@nrwl/cli``, the two version numbers should be always the
same.