fix(testing): issue passing coverageRepoters commandline
In Jest coverageReporters are an array of strings (https://jestjs.io/docs/en/configuration#coveragereporters-arraystring). In here it is an optional string.
When passing a string value for the coverageReporters to the command line, it results into the following error:

Failed to write coverage reports:
ERROR: TypeError: coverageReporters.forEach is not a function
STACK: TypeError: coverageReporters.forEach is not a function