fix(core): fix module resolution in project locator
On windows, there is a bug in project module resolution, cause by path separator.  TypeScript resolveModuleName will always return a posix normalized path, so our appRootPath needs to be normalized for string replacement.

I came across the issue when trying to run dep-graph, but other areas (ex: affected:*) which rely on TargetProjectLocator are going to be impacted.