docs(angular): fixed link to page nx and cli
When introducing the tool to a friend, I clicked on the link "learn more about nx cli and angular
cli", which resulted in a blank page;  there was `%60` in the URL. After decodeURI, I thought it
would have the "grave accent" (used in template string) that should not be there. So, I removed it.