feat(core): adds ability for outputs to be files and not just folders
Makes the cache put code slightly more flexible, which allows for specified outputs to be not
just
folder but also files. This allows for greater flexibility of control when it comes to which
files
you want Nx to cache, so you could exclude sub folders of your dist. This is useful so that
Nx
doesn't spend time copying sometimes complex folder structures to cache (i.e. node_modules in
the
case of using npm link).