chore(repo): update resources in node modules
Part 2 of PR #4235 - carrying over the same changes to this file.

Description of changes: I removed the list of individual talks, presentations, and podcasts (maintaining the doc with every new talk is a lot of work, and the list was getting very long). We'll maintain a YouTube playlist with all that info. The playlist and Office Hours playlist are now linked in the "Additional Resources" section (formerly known as "Misc").