fix: adding support for production-only environment variables on build
Currently when building a @nrwl/react project, all environment variables are loaded for any type of build: production or otherwise. This commit successfully determines whether or not the user is trying to run a task in a production configuration, and loads only the appropriate environment variables (excluding all `.local.env` files) in the build process.