fix(core): add Hasher test to point out the implicit deps bug
The list of projects in the `nx.json` implicitDeps is falsely ignored.
This is in conflict with [specification](https://github.com/nrwl/nx/blob/master/docs/shared/affected.md#when-nx-cant-understand-your-repository):

```
The implicitDependencies map is used to define what projects are affected by global files.
(...)
You can replace * with an explicit list of projects.
```