feat(core): delete test conflicting with the deps globing
In versions prior 12.0.0, any other string than `*` as an implicit dependency
would cause stack overflow. This issue was fixed in commit a300e76, which
enforces that only `*` or array can be specifier for the implicit dependencies.

We conclude that supporting more general glob patterns is not a risk from
backward compatibility standpoint.