fix(core): support hidden files in assets (#6298)
The transformation of `AssetGlob` objects into `FileInputOutput` objects does not include hidden files. The glob option `dot: true` is required to include hidden files. The commit extends the AssetGlob interface with the optional property `dot`. The glob check will include hidden files when the property `dot` is set to true.

Signed-off-by: Merzough Münker <mmuenker@digitaix.com>