Don't disable chalk when FORCE_COLOR is set
My lint output in Jenkins wasn't colored, which is pretty normal because
a lot of apps when running in Jenkins don't think they have a
color-capable terminal to write to, but I expected that setting the
`FORCE_COLOR` environment variable would work to override this, but I
was surprised to find that it didn't help. Looking at the Nx code, I see
that chalk is turned off whenever `CI` is true, which seems like it
should be refined a bit.