fix(angular): fixing move generator to update all import path
There is a known issue with RegExp that exists for many years on test methods, that causes test result to depend on previous calls to test method.
So if the first one matches, the second time the test is run, it will not match even if it should.
By reseting the last index of the regex, the test returns the correct result every time.
This issue was causing import path to be updated only on approxymately half of files.