fix(testing): update tsconfig migration to not warn when no file found (#7689)
* fix(testing): update tsconfig migration to not warn when no file found

use tsconfig.json references to find what other configs need to be updated. if non when we just exit

* fix(testing): update tests for migration changes

* fix(testing): include any spec. file names in pattern match

accounting for _spec. file names when generating 'test' pattern equivalents

* fix(testing): remove relance on tsconfig.references, walk for all tsconfigs to update

tsconfig.references is not always present. instead visit all files and check if they are tsconfig
json file. if so we update the include/exclude properties where a spec. pattern is found. if the
property isn't set then we do not update it.

* fix(testing): stream line update with devkit