fix(storybook): added outDir to tsconfig compilerOptions in Angular Storybook configuration
When omitting outDir in the tsconfig of the Storybook configuration will cause the build to fail, if
allowJS is set to true.
The compiled TypeScript files will be included again in the build as JS
files. Changed the template to set the outDir to the dist folder of the repository root.

ISSUES CLOSED: #7825