Update Jenkinsfile example
While our team was trying to migrate to Nx, we found out this issue: https://github.com/nrwl/nx/issues/2170
The documentation doesn't mention about this behaviour but in our case setting the --base to origin/{current_branch} fixed it. I've set the env variable accordingly in the md file.