fix(node): wait for subprocess to exit when --watch=false in node execute executor
Fixes #7031.

Before this fix, the executor would exit immediately and let the node subprocess continue in the background,
if watch was set to false.
This commit attempts to fix this problem, by adding event listeners on the subprocess in the node/execute
executor if watch is false, such that the nx process continues simply (a)waits until the subprocess exists
(successfully or not).

Error handling has also been added such that the executor fails if a non-successful status code is returned by the program