fix(storybook): look for config dir/files relative to the project dir
When specifying a config file/dir path in the executor options.config object, the statSync() call
fails if the user does not run the `nx ...` command from the root of the repository/workspace.
This is confusing and inconsistent with the way that other executors work, like @nrwl/jest.

ISSUES CLOSED: #9115