fix(misc): fix format:check not checking all affected files
While running format:write and format:check commands, a list of affected files is split to chunks of 50 files, that was originally meant as compatibility optimization for Windows users (see #2362).
This causes early exits in cases if there are also incorrectly formatted files in both 1st and following chunks.

fixes #9151