feat(js): infer tslib as dependency when using importHelpers
When using "importHelpers" and building with "tsc", the built bundle expects to be supplied the
"tslib" package.

Currently, it's up to developers to add "tslib" as a dependency to those
libraries' package.json files.

This PR makes the "tsc" execute infer that "importHelpers" is used,
and subsequently add "tslib" as a dependency to the generated package.json

ISSUES CLOSED: #9343