fix(core): local plugins should resolve in package-manager workspaces
In npm/pnpm/yarn workspaces, packages apparently end up symlinked to node_modules. This was causing local plugin resolution to fail, since the fallback call was never hit. The fix is to check for a local plugin *first*, and then if not found to fall back on the installed plugin.

This has the disadvantage of it not being possible to hit the installed version of a plugin's executors / generators, since the local plugin resolution will always kick in. I'm not super sure if that is worth fixing, but its  something to consider.

Fixes #9823