cleanup(core): provide better cleanup api for createTempNpmDirectory
Implement a suggestion by @agentender from a previous commit that provides a cleanup function for a
temp directory, alongside the directory path itself.

This is done due to the fact that the method seems to be of use to users of NX and not only NX itself,
and it's fitting that we make the burden of cleaning-up the temp directory easier, with less boilerplate to copy.