fix(core): fix for @nrwl/nx/enforce-module-boundaries linting rule to be slow on Windows
The eslint nx-enforce-module-boundaries rule caches the NX graph during a full eslint run. Based on
how the eslint rule is started (ea is het started by the user from a terminal or is het started for
1 file by the IDE), the cache will
be rebuilt or not.
To determine if eslint is started via a
terminal, NX looks into the process argv to find out. On Windows the path in process argv is with
backslashes instead of forward slashes. The regex did not took into account back slashes.

ISSUES CLOSED: #11250