docs(misc): updated AST playground to TSQuery (#11295)
* docs(devkit): updated the astexplorer link to point to the tsquery demo.

The docs were pointing to the AST Explorer demo but it mentioned using the TSQuery npm package. So I
changed the AST Explorer link to the TSQuery Playground.

* docs(misc): updated AST playground to TSQuery

The docs were linking to the AST Explorer, which is not the npm package referenced in the generator
documentation (as AST Explorer does not support parsing TypeScript).

* docs(misc): updated AST playground to TSQuery

The docs were linking to the AST Explorer, which is not the npm package referenced in the generator
documentation (as AST Explorer does not support parsing TypeScript).

* docs(misc): updated AST playground to TSQuery

The docs were linking to the AST Explorer, which is not the npm package referenced in the generator
documentation (as AST Explorer does not support parsing TypeScript).