fix(core): correctly handle overrides of watch flag for caching
Previously if the --watch flag were being overridden, when attempting to determine if a given task is long-running, we would perform a !!value check against the override. Since overrides are passed through as strings, if a --watch override was provided at all, the task would be marked as long running -- even if the override was disabling watch mode.

This has a side effect of the Nx Cloud runner not persisting task outputs to the remote cache, causing any task with a --watch override running in DTE to fail.