feat(core): use node_modules/.cache/nx as default for projectGraphCacheDirectory
The default value for projectGraphCacheDirectory is
changed to `node_modules/.cache/nx`.
The
rationale for this is that if someone moves the
nx cache of artefacts by setting
`NX_CACHE_DIRECTORY`
or setting `cacheDirectory` in `nx.json` to some location
that will be shared
among multiple worktrees/repos then
it is undesirable to have daemon and project graph cache
shared
too.
This is a BREAKING change as projectGraphCacheDirectory
would previously move with the achanges
to `cacheDirectory`
and/or `NX_CACHE_DIRECTORY`.
It is however not a breaking change that is
expected to
affect anyone negatively and in the unlikely case we have
environment
NX_PROJECTNX_PROJECT_GRAPH_CACHE_DIRECTORY

BREAKING CHANGE:
The default of projectGraphCacheDirectory no longer follows the location of cacheDirectory