docs(nxdev): refactor docs sources generation and app consumption
It refactors the way the NextJS application is consuming the generated data from the documentation and the way the said data is generated.

This introduces multiple modifications to the way the documentation is generated, what is generated and how it is consumed by the application.

The packages data introduce the following architecture:
```
/{package-name}
  - documents/ (markdown)
  - executors/ (json)
  - generators/ (json)
```
This allows us to directly show the generation results of what is contained in the `map.json`.

We are now generating a manifest for each type of data we encounter in the application:
- packages
- documents (recipes, cloud, nx)
- menus
- tags

Each type has it own type shared by the generator and the application for stronger compliance and compatibility. These manifets contain the data ready to be consumed, meaning all the custom logic has been applied by the generator.

As an example, documents associated to a specific packages referenced in `additional-api-references` are traited exactly the same as the ones for Cloud, Nx or even the recipes. They share the same interface and the same faetures, like the tagging system.

The application has been heavily updated in two aspects:
- The routes are now a mix between file based and dynamic routing to answer the custom URL scheme we are following as well as simplifying the template and data handling.
- Only the relevant data is loaded when it is needed. On development, the data load is plit into multiple endpoint. We were loading everything at once previously.
- Data handling is simpler, with less custom business logic and only 2 main interfaces to follow: `DocumentMetadata` and `PackageMetadata`.
- A new `DocumentsApi` and `PackagesApi` that allows us to get a cleaner set of URLs, removing the duplication and other different path errors, that produced URLs that were not supported.
- A new `TagsApi` that works with the `DocumentsApi` to easily find
  releated document thoughout all the available data of type
  `DocumentMetadata`.