feat(core): add options to disable loading dotenv files
Add task runner option 'loadDotEnvFiles' to the default task runner.
When this option is false, no .env file as described in the env file
hierarchy of the Nx docs will be loaded. When this option is explicitly
set to true or not set at all, then the .env files will be loaded
normally.

Add option 'loadDotEnvFile' to run-commands.
When this option is false, or the task runner's 'loadDotEnvFiles' is
explicitly set to false, then the ".env" file in the CWD will not be
loaded.
When this option is explicitly set its behaviour will override the
option 'loadDotEnvFiles' from the task runner.
This option will be ignored if the 'envFile' option is present.
When this option is explicitly set to true or not set at all, then
by default the ".env" file from the CWD will be loaded normally.