fix(detox): Update rn and expo buildTargets
Previously, the React Native and Expo test targets were not being created with the correct buildTarget for the platforms iOS and Android. This caused issues when running e2e tests for Android.

To fix this issue, I updated the React Native and Expo test targets to ensure that the correct buildTarget is created for each platform. This ensures that the creation of project.json is accurate and reliable.