fix(expo): Add @testing-library/jest-native to tsconfig compiler options for type checking
Previously, our project was not properly recognizing the types for the "@testing-library/jest-native" package, resulting in type errors  in our test files within VSCode. To fix this issue, I added the entry  "@testing-library/jest-native" to the "types" array in the tsconfig  file. By doing this, Visual Studio Code (VSCode) is now able to  pick up the types for the package, allowing for proper type checking  and no more type errors in our test files.