fix(nextjs): metadata implementation nextjs root layout
The layout.tsx file component generated in a new NextJS application with appDir enabled is using
both the Head component from next/head and implementing the Metadata API, which leads to a warning
message. This commit removes the redundant Head component to optimize the generated code and resolve
the warning message.

closed 16832