fix(nextjs): error caused by undefined 'serverActions'
This commit addresses a bug introduced in the PR https://github.com/nrwl/nx/pull/16819, which caused
a TypeError when running NextJS 13.4 with the '_nextConfig.experimental' object missing the
'serverActions' property. The fix involves adding checks to ensure that both the
'_nextConfig.experimental' object and 'serverActions' property are defined before attempting to
access them. If 'serverActions' is undefined, the code defaults to using 'next'.

closed 16845