fix(core): do not mix UNIX and performance API timestamps in profiles
`TaskProfilingLifeCycle` calculates profile timestamps and durations based
on a mix of performance timestamps (coming from `performance.now()`) and
UNIX ones (coming from tasks).

The problem comes when the two types of values are mixed - e.g.
`perfStart` comes from the task and `perfEnd` comes from the performance
API, or vice-versa. Since performance timestamps are relative to the
process start and UNIX timestamps are relative to the UNIX epoch, such
cases would result in wrong calculations - e.g. an enormous negative
value if `perfStart` is a UNIX timestamps and `perfEnd` is a performance
timestamp.

Closes #18388