fix(misc): pass full path to tsconfig when registering transpiler
Currently when invoking registerTsProject you are meant to pass two arguments:
- The root directory that arg #2 is relative to
- The name of the tsconfig

This would look something like: `registerTsProject('path/to/my-project', 'tsconfig.app.json')`. There are a few places in the codebase where I found that the first path was not to this root, and we never used that arg without also joining on the tsconfig name at the end.

This updates the helper function to take a singular arg, the path to the tsconfig. The original signature is deprecated and will be removed in v18, since we had a docs page that referenced it. We could add a migration for this if anyone thinks its worth it, but I think the deprecation notice should suffice.

The bad method calls surfaced as a warning that tsconfig-paths wasn't installed, which was clearly wrong. This was because we tried to invoke tsconfig paths on a bad path (it looked something like 'path/to/my-project/module-federation.config.ts/tsconfig.json'). This PR also updates the relevant places that called the helper to use the new signature