feat(core): allow marking targets hidden
Sometimes it doesn't make sense for a target to show up in nx console or nx show. The main case for this currently is `nx-release-publish`, which should never be manually invoked. Its worth generalizing this though, as I can see it making sense for other targets if they shouldn't be manually invoked under normal circumstances. For example, if setting up code generation as a dependency of build, it may make sense to hide the codegen target since it should always be called during builds