feat(vite): allow passing path to custom tsconfig file when skipTypeCheck is false
When skipTypeCheck is false and no tsconfig file exists in projectRoot, it returns undefined which
causes `join(opts.workspaceRoot, opts.tsconfig)` to fail with `The "path" argument must be of type
string. Received undefined`.
Allow user to pass custom tsconfig file location which can be in parent/child directory of
projectRoot