docs(core): fix dependsOn snippet in project configuration reference (#23668)
I'm not totally sure, but by reading the docs, the intent of the first
example is to say that the test target depends of the build target.
Otherwise, it would mean that the build target depends of the build
target, which does not make sense. Or did I misunderstood something in
the doc ?

Co-authored-by: U812320 <francoisjulien.chabbey@mobi.ch>
(cherry picked from commit 32030435b74b236b5fddafbc1e42513fb0a750e8)