feat: improve error handling for npm publish
This change improves error handling for npm publish function. When there is any error that is not handled by the catch clause, it is throwing an "cannot read summary of undefined", because the stdoutData.error is undefined. With this change, it is going to correctly throw the root error that was the cause of the problem.